requirements.txt
playwright
dotenv
pyyaml
customtkinter